+++ tests/engine/recommendation_engine_test.py
@@ -1,24 +1,10 @@
 import os
 import unittest
 
-import pandas as pd
-
 from mediabridge.db.connect import connect_to_mongo
-from mediabridge.definitions import DATA_DIR, OUTPUT_DIR
+from mediabridge.definitions import OUTPUT_DIR
+from mediabridge.engine.engine_etl import etl_mongo_movie_titles
 from mediabridge.engine.recommendation_engine import RecommendationEngine
-
-
-def etl_mongo_movie_titles() -> None:
-    df = pd.read_csv(DATA_DIR / "movie_titles.csv")
-    df["netflix_id"] = df.netflix_id.astype(str)
-    rows = df.to_dict(orient="records")
-    print(rows)
-
-    client = connect_to_mongo()
-    movies_collection = client["movies"]
-
-    movies_collection.insert_many(rows)
-
 
 # In a github Actions CI environment, we will be "disabled".
 _enabled = bool(os.getenv("MONGODB_URI"))
